// web/src/api/generated/model/workflowTaskDefSchema.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.6.10
 */
import type { WorkflowTaskDefSchemaConfig } from './workflowTaskDefSchemaConfig';

export interface WorkflowTaskDefSchema {
  id: string;
  type: string;
  config?: WorkflowTaskDefSchemaConfig;
  upstream?: string[];
  result_key?: string | null;
}